+++ datadog-checks-base/datadog_checks/stubs/__init__.py
@@ -1,8 +1,0 @@
-# (C) Datadog, Inc. 2018
-# All rights reserved
-# Licensed under a 3-clause BSD style license (see LICENSE)
-from .aggregator import aggregator
-
-__all__ = [
-    'aggregator', 'datadog_agent'
-]