+++ datadog-checks-tests-helper/datadog_test_libs/win/pdh_mocks.py
@@ -5,7 +5,6 @@
 import os
 import pytest
 import mock
-import re
 import string
 import _winreg
 from collections import defaultdict
@@ -38,7 +37,28 @@ def pdh_mocks_fixture():
     pdhgetformattedcountervalue.stop()
     pdhcollectquerydata.stop()
 
-def initialize_pdh_tests(regdata = None, counterdata = None):
+
+@pytest.fixture
+def pdh_mocks_fixture_bad_perf_strings():
+    regqueryvalueex = mock.patch('_winreg.QueryValueEx', mock_QueryValueExWithRaise)
+    pdhlookupbyindex = mock.patch('win32pdh.LookupPerfNameByIndex', mock_LookupPerfNameByIndex)
+    pdhenumobjectitems = mock.patch('win32pdh.EnumObjectItems', mock_EnumObjectItems)
+    pdhmakecounterpath = mock.patch('win32pdh.MakeCounterPath', mock_MakeCounterPath)
+    pdhaddcounter = mock.patch('win32pdh.AddCounter', mock_AddCounter)
+    pdhgetformattedcountervalue = mock.patch('win32pdh.GetFormattedCounterValue', mock_GetFormattedCounterValue)
+    pdhcollectquerydata = mock.patch('win32pdh.CollectQueryData', mock_CollectQueryData)
+
+    yield regqueryvalueex.start(), pdhlookupbyindex.start(), pdhenumobjectitems.start(),  pdhmakecounterpath.start(), pdhaddcounter.start(), pdhgetformattedcountervalue.start(), pdhcollectquerydata.start()
+
+    regqueryvalueex.stop()
+    pdhlookupbyindex.stop()
+    pdhenumobjectitems.stop()
+    pdhmakecounterpath.stop()
+    pdhaddcounter.stop()
+    pdhgetformattedcountervalue.stop()
+    pdhcollectquerydata.stop()
+
+def initialize_pdh_tests(lang = None):
     """
     initialize_pdh_tests
 
@@ -50,42 +70,59 @@ def initialize_pdh_tests(regdata = None, counterdata = None):
 
     Can be re-initialized with new values if desired
 
-    Parameters:
-    regdata     Filename of fixture that contains the registry
-                data.  If not specified, the provided data is
-                used.
-
-    counterdata Filename which contains the available installed
-                counters
-
     """
-    if regdata is None:
-        regdata = os.path.join(HERE, 'fixtures', 'counter_indexes_en.txt')
-    if counterdata is None:
-        counterdata = os.path.join(HERE, 'fixtures', 'allcounters.txt')
+    en_us_reg_file = "counter_indexes_en-us.txt"
+    if lang is None:
+        local_reg_file = en_us_reg_file
+        file_base = "allcounters_en-us.txt"
+    else:
+        local_reg_file = "counter_indexes_%s.txt" % lang
+        file_base = "allcounters_%s.txt" % lang
 
-    load_registry_values(regdata)
+    en_us_reg_file = os.path.join(HERE, 'fixtures', en_us_reg_file)
+    local_reg_file = os.path.join(HERE, 'fixtures', local_reg_file)
+
+    counterdata = os.path.join(HERE, 'fixtures', file_base)
+
+    eng_array, eng_indexes = load_registry_values(en_us_reg_file)
+    global index_array
+    global counters_index
+    index_array = eng_array
+    if lang is None:
+        counters_index = eng_indexes
+    else:
+        loc_array, loc_indexes = load_registry_values(local_reg_file)
+        counters_index = loc_indexes
+
     read_available_counters(counterdata)
 
 def read_available_counters(fname):
     counters_by_class.clear()
     instances_by_class.clear()
-    regex = r'(\\[^(]+)(?:\(([^)]+)\))?(\\[^(]+)'
-    compiled_regex = re.compile(regex)
 
     with open(fname) as f:
         for line in f:
-            line = line.rstrip()
-            clss, inst, counter = [x if not x else x.lstrip('\\') for x in compiled_regex.search(line).groups()]
+            line = line.strip()
+            if line.startswith("#"):
+                continue
+
+            try:
+                clss, inst, counter = line.split('!')
+                clss = clss.strip()
+                inst = inst.strip()
+                counter = counter.strip()
+            except:
+                print line
+                raise
+
             counters_by_class[clss].add(counter)
             if inst:
                 instances_by_class[clss].add(inst)
 
 
 def load_registry_values(fname):
-    global index_array
-    index_array = []
-    counters_index.clear()
+    idx_array = []
+    ctr_index = defaultdict(list)
 
     with open(fname) as f:
         linecount = 0
@@ -96,22 +133,34 @@ def load_registry_values(fname):
                     break
                 line = " "
 
-            index_array.append(line)
+            idx_array.append(line)
             linecount += 1
 
     idx = 0
-    idx_max = len(index_array)
+    idx_max = len(idx_array)
     while idx < idx_max:
-        counters_index[int(index_array[idx])] = index_array[idx+1]
+        ctr_index[int(idx_array[idx])] = idx_array[idx+1]
         idx += 2
+    return idx_array, ctr_index
 
 def mock_EnumObjectItems(reserved, machine_name, clss, detail):
     ctrs = list(counters_by_class[clss])
     insts = list(instances_by_class[clss])
     return ctrs, insts
 
-def mock_MakeCounterPath(machine, clss, inst, parent, iindex, cname):
-    ## do some sanity checking
+def mock_MakeCounterPath(arglist):
+    # win32pdh.MakeCounterPath takes as it's first argument a list of parameters
+    # to pass to the underlying C API.  Providing the unused parameters (commented
+    # out) as "documentation" for the params; especially in the event they're
+    # needed in future improvements
+
+    # machine = arglist[0] # remote machine name
+    clss = arglist[1]
+    inst = arglist[2]
+    # parent = arglist[3]
+    # iindex = arglist[4]
+    cname = arglist[5]
+    # do some sanity checking
     if clss not in counters_by_class:
         raise AttributeError("class name {} is invalid".format(clss))
     if inst and inst not in instances_by_class[clss]:
@@ -126,6 +175,8 @@ def mock_MakeCounterPath(machine, clss, inst, parent, iindex, cname):
     return p
 
 def mock_AddCounter(h, path):
+    if path is None or len(path) == 0:
+        raise AttributeError("Invalid path")
     return path
 
 def mock_GetFormattedCounterValue(h, p):
@@ -134,10 +185,11 @@ def mock_GetFormattedCounterValue(h, p):
 def mock_CollectQueryData(h):
     return True
 
+def mock_LookupPerfNameByIndex(machine_name, ndx):
+    return counters_index[ndx]
 
 def mock_QueryValueEx(*args, **kwargs):
     return (index_array, _winreg.REG_SZ)
 
-
-def mock_LookupPerfNameByIndex(machine_name, ndx):
-    return counters_index[ndx]
+def mock_QueryValueExWithRaise(*args, **kwargs):
+    raise WindowsError