+++ datadog_checks_base/datadog_checks/checks/__init__.py
@@ -1,0 +1,13 @@
+# (C) Datadog, Inc. 2018
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+
+try:
+    # Agent5 compatibility layer
+    from checks import AgentCheck
+except ImportError:
+    from .base import AgentCheck
+
+__all__ = [
+    'AgentCheck',
+]