+++ datadog_checks_base/datadog_checks/stubs/_util.py
@@ -1,0 +1,10 @@
+# (C) Datadog, Inc. 2018
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+
+
+class SubprocessOutputEmptyError(Exception):
+    pass
+
+def subprocess_output(*args, **kwargs):
+    pass