+++ statsd/CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG - statsd
 
+1.1.0 / Unreleased
+==================
+
+### Changes
+
+* [FEATURE] Add timeout configuration.
+
 1.0.0 / 2017-03-22
 ==================